Finish the state dump of tendermint.
